+++ src/components/Header.jsx
@@ -1,25 +1,13 @@
 import { HeaderBack } from "../css/other"
+import { HeaderTime } from "./header-component/Headertime"
+import { HeaderMessage } from "./header-component/HeaderMessage"
 
 export const Header = () => {
-
-  const getTime = () => {
-    let today = new Date();
-    let minute = today.getMinutes()
-    let hour = today.getHours();
-    let date = today.getDate();
-    let month = today.getMonth()+1;
-    var minuteWithZero = ( '000' + minute ).slice( -2 );
-
-    return `${month}.${date}   ${hour}:${minuteWithZero}`
-  }
-  
   return (
     <HeaderBack className="md:h-60 h-40 font-bold font-sans">
-      <h1 className="sm:text-4xl md:ml-20 ml-2 text-2xl">
-        HELLO!<br/>
-        What’s your plan today?
-        <p className="sm:text-2xl text-xl mt-2 sm:mt-4">DATE  :{getTime()}</p>
-      </h1>
+      <h1 className="sm:text-4xl text-2xl">HELLO!</h1>
+      <HeaderMessage/>
+      <p className="sm:text-2xl text-xl mt-2 sm:mt-4">DATE  :{HeaderTime()}</p>
     </HeaderBack>
   )
 }
